module2-task1: правки по ревью
