+++ src/enums/MarkCardEnum.ts
@@ -1,0 +1,3 @@
+export enum MarkCard {
+  Premium = 'Premium'
+}
