+++ src/enums/TypeCardEnum.ts
@@ -1,0 +1,4 @@
+export enum TypeCard {
+  Apartment = 'Apartment',
+  Room = 'Room'
+}
