--- src/props/FavoritePlace.ts
@@ -1,8 +1,0 @@
-import CardProps from './FavoriteCardProps';
-
-type FavoritePlaceProps = {
-  name: string;
-  cards: CardProps[];
-}
-
-export default FavoritePlaceProps;
