--- src/props/SettingProps.ts
@@ -1,5 +1,0 @@
-type SettingProps = {
-    offersCount: number;
-}
-
-export default SettingProps;
