+++ src/types/FavoritePlace.ts
@@ -1,0 +1,8 @@
+import Card from './Card';
+
+type FavoritePlace = {
+  name: string;
+  cards: Card[];
+}
+
+export default FavoritePlace;
