+++ src/types/Setting.ts
@@ -1,0 +1,5 @@
+type Setting = {
+    offersCount: number;
+}
+
+export default Setting;
